src/requirements.txt
beautifulsoup4
bs4
html5lib
mechanize
python-dotenv
selenium
six
soupsieve
urllib3
webencodings
peewee
